fix behavior for multi-turn revolute joints
Previously, the behavior of revolute joints was unexpected when
the joint limits extended beyond [-pi, pi] due to unwrapping
happening in a few places in the angles package, which was called
from the joint position controller. This PR checks to see if the
range of motion extends beyond [-pi, pi] and, if so, goes directly
to the (clamped) target, rather than trying to unwrap the current
or target angles to find the shortest path.