add test case that fails without e53b364
See pull request for fix/jerk_on_stop for detailed description.

This commit adds a delay subscriber to rrbot to emulate a hardware controller with dead time.
The jump backwards in the actual position is observed in joint_trajectory_controller_test by
extending the stateCB.
The new testCase fails, if the velocity of the actual position is negative although motion should only be forwards.