[effort_controllers] JointGroupPositionController: Check output of angles::shortest_angular_distance_with_large_limits
The current implementation does not check the output of
angles::shortest_angular_distance_with_large_limits which indicates if
the returned distance is an unspecified (and potentially wrong) case.

This commit checks the return statement and adds a detailed warning.

Secondly, the implementation in
angles::shortest_angular_distance_with_large_limits is only valid if the
`from` argument is within the range given by the upper and lower joint
limit. As thus, the current_position is now enforced/clamped to the
joint limits.